requirements.txt
# Dependencias para el proyecto
PyQt5>=5.15.0
mysql-connector-python>=8.0.0
websockets>=10.0
asyncio

# Para deployment en servicios cloud
gunicorn>=20.0.0
python-dotenv>=0.19.0

# Para seguridad
cryptography>=3.0.0

# Para manejo de archivos estáticos (opcional)
flask>=2.0.0
flask-cors>=3.0.0
